feat(metasploit): add module metadata filters
